Ship RBD OCF RA /w charm.
The rbd OCF was not included with the ceph-resource-agents package
until quantal. There is currently not an easy way to install from
another repository as it will also upgrade other ceph package in the
principle service.  As a temporary alternative, ship the rbd OCF
/w the charm for 12.04.

Also improve logging for cases where principle has not passed minimum
required corosync config.