Change the default cluster_count to 3
Most production deployments involve 3 nodes for high availability.
This change modifies the default cluster_count to 3 in order to
align with the typical deployment scenario. Previously, the default
cluster count was 2 which is often unmodified and then leads to
deployment failures due to a mismatch in the nodelist defined in
each node's corosync.conf file.

Change-Id: I0799d8b880ecdb9c933d0361e7dc843b68fc5c82