Fix pacemaker down crm infinite loop
On corosync restart, corosync may take longer than a minute to come
up. The systemd start script times out too soon. Then pacemaker which
is dependent on corosync is immediatly started and fails as corosync
is still in the process of starting.

Subsequently the charm would run crm node list to validate pacemaker.
This would become an infinite loop.

This change adds longer timeout values for systemd scripts and adds
better error handling and communication to the end user.

Change-Id: I7c3d018a03fddfb1f6bfd91fd7aeed4b13879e45
Partial-Bug: #1654403