Add pacemaker authkey
To work with pacemaker remotes all pacemaker nodes (including the
remotes) need to share a common key in the same way that corosync
does. This change allows a user to set a pacemaker key via config
in the same way as corosync. If the pacemaker key value is unset
then the corosync key is user.

Change-Id: I75247e7f3af29fc0907a94ae8e1678bdb9ee64e2