Stop HA services accross units for series upgrade
Stop HA services accross all units of an application when doing a
series upgrade to avoid the situation where the cluster has some
nodes on LTS N-1 and some on LTS N.

1) In the 'pre-series-upgrade' send a notification to peers informing
   them that the unit is doing a series upgrade and to which Ubuntu
   version.
2) Peers receive notification. If they are on a later Ubuntu version
   than the one in the notification then they do nothing. Otherwise
   they shutdown corosync and pacemaker and add an entry to the local
   kv store with waiting-unit-upgrade=True.
3) In the 'post-series-upgrade' hook the notification is removed from
   the peer relation. waiting-unit-upgrade is set to False and
   corosync and pacemaker are started.

The result of this is that when the first unit in the cluster starts
a series upgrade all cluster services are shutdown across all units.
They then rejoin the cluster one at a time when they have been
upgraded to the new version.

I added the waiting-unit-upgrade key to deal with the situation where
the first node clears the notification after it has successfully
upgraded, with out the waiting-unit-upgrade the peers would not know
they were in a mixed Ubuntu version cluster.

Change-Id: Id9167534e8933312c561a6acba40399bca437706
Closes-Bug: 1859150