Add focal-ussuri and bionic-ussuri bundle
This patch adds a focal-ussuri and bionic-ussuri bundles to the tests
for the charm.  Note that this requires a modified zaza (to use the
force_deploy option).

Note that the func-test change below is to fix the zaza test that
assumes that all subordinates are hacluster; in focal, there is also a
mysql-router subordinate.

func-test-pr: https://github.com/openstack-charmers/zaza-openstack-tests/pull/241
Change-Id: I17b9617917a4e0e14817c97023a4676db57937c0