Add docs key and point at Discourse
Add the 'docs' key and point it at a Discourse topic
previously populated with the charm's README contents.

When the new charm revision is released to the Charmhub,
this Discourse-based content will be displayed there. In
the absense of the this new key, the Charmhub's default
behaviour is to display the value of the charm's
'description' key.

Depends-On: https://review.opendev.org/c/openstack/charm-hacluster/+/895818
Change-Id: Icf59d65f7938a9a72a0dca87dfbcf13118fabea1